refactor: move exceptions handling to middleware
