+++ lib/anycable/grpc/handler.rb
@@ -11,41 +11,14 @@ module AnyCable
       # Handle connection request from WebSocket server
       def connect(request, _unused_call)
         AnyCable.rpc_handler.handle(:connect, request)
-      rescue => exp
-        notify_exception(exp, :connect, request)
-
-        AnyCable::ConnectionResponse.new(
-          status: AnyCable::Status::ERROR,
-          error_msg: exp.message
-        )
       end
 
       def disconnect(request, _unused_call)
         AnyCable.rpc_handler.handle(:disconnect, request)
-      rescue => exp
-        notify_exception(exp, :disconnect, request)
-
-        AnyCable::DisconnectResponse.new(
-          status: AnyCable::Status::ERROR,
-          error_msg: exp.message
-        )
       end
 
       def command(request, _unused_call)
         AnyCable.rpc_handler.handle(:command, request)
-      rescue => exp
-        notify_exception(exp, :command, request)
-
-        AnyCable::CommandResponse.new(
-          status: AnyCable::Status::ERROR,
-          error_msg: exp.message
-        )
-      end
-
-      private
-
-      def notify_exception(exp, method_name, message)
-        AnyCable::ExceptionsHandling.notify(exp, method_name.to_s, message.to_h)
       end
     end
   end
